part3

+++ OGP1516-Hillbillies/src-provided/hillbillies/part1/facade/Facade.java
@@ -23,7 +23,7 @@ public class Facade implements IFacade {
 	}
 
 	@Override
-	public long[] getCubeCoordinate(Unit unit) throws ModelException {
+	public int[] getCubeCoordinate(Unit unit) throws ModelException {
 
 		return unit.getOccupiedGameCube(unit.getPoistion());
 	}
